* master: (863 commits)
  Revert "mpich: use latest mpich tarball with gcov but use '--enable-coverage' mpich configure option."
  MatCreateTranspose: fix for lazy block size
  mpich: use latest mpich tarball with gcov but use '--enable-coverage' mpich configure option.
  configure: add FC_DEFINE_FLAG to petscvariables
  added a citation for pounders, Stefan finally came through
  update to latest netcdf fixes bug that did not allow building on recent Mac OS X
  DMPlex: Fixes for CGNS - Fixed loop bounds since CGNS uses closed intervals - Bug with mixed cell types - Typo with coordinates
  final updates from nightly tests
  fix for countcfunction
  removed outdated docs and comments on MatStructure usage removed SAME_PRECONDITIONER since no longer used removed unneed use of MatStructure in Schur complement matrix
  removed MatStructure arguments from Tao
  fixes for Sundials and fun3d cannot use MPIU_SUM for integer arrays
  inplace ILU factorization must update state of inplace matrix to prevent infinite recursion of updating PC
  fixes for BDDC by the way not having BDDC always in the testing is going to lead to trouble
  the compute Jacobian functions for SNES and TS now do not take a MatStructure flag
  KSPSetOperators() no longer has the MatStructure argument. The Mat objects now track that information themselves. Use KPS/PCSetReusePreconditioner() to prevent the recomputation of the preconditioner if the operator changed in the way that SAME_PRECONDITIONER did with KSPSetOperators()
  tracks when Mat's have a change in nonzero structure this information can be used by PC to determine how preconditioners must be rebuilt currently only does an error check if the user said same nonzero pattern but actually changed the pattern
  completed cleanup of removal of matrix pointers to compute jacobian routines
  makefile: Added 'make includegraph' to create an inclusion dependency graph.
  Contrib: Updated include-graph.sh script to ignore *.hh files
  ...

Conflicts:
	src/dm/dt/interface/dt.c

+++ src/snes/examples/tutorials/ex25.c
@@ -62,7 +62,7 @@ int main(int argc,char **argv)
   litspit = ((PetscReal)lits)/((PetscReal)its);
   ierr    = PetscPrintf(PETSC_COMM_WORLD,"Number of SNES iterations = %D\n",its);CHKERRQ(ierr);
   ierr    = PetscPrintf(PETSC_COMM_WORLD,"Number of Linear iterations = %D\n",lits);CHKERRQ(ierr);
-  ierr    = PetscPrintf(PETSC_COMM_WORLD,"Average Linear its / SNES = %e\n",litspit);CHKERRQ(ierr);
+  ierr    = PetscPrintf(PETSC_COMM_WORLD,"Average Linear its / SNES = %e\n",(double)litspit);CHKERRQ(ierr);
 
   ierr = SNESDestroy(&snes);CHKERRQ(ierr);
   ierr = PetscFinalize();
